+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 Newfeatures
 
 - `Identifier search parameter <http://www.hl7.org/fhir/search.html#token>`_ is active now.
+- Array of Reference query support (for example `basedOn` (list of reference) ) is active now.
 
 
 1.0.0b7 (2018-08-10)